LayoutTests/http/tests/media/video-useragent-expected.txt


Tests that the media player sends the WebKit User Agent string when requesting a media file.
EVENT(canplay)
END OF TEST
